Added missing transaction
#rb none
#rnx

[CL 15265269 by Jamie Dale in ue5-main branch]

+++ Engine/Source/Editor/UnrealEd/Private/UnrealEdSrv.cpp
@@ -2498,8 +2498,12 @@ bool UUnrealEdEngine::Exec_Actor( UWorld* InWorld, const TCHAR* Str, FOutputDevi
 
 				TStrongObjectPtr<UTypedElementList> ElementsToDelete(UTypedElementRegistry::GetInstance()->CreateElementList());
 				CommonActions->GetSelectedElementsToDelete(SelectionSet, ElementsToDelete.Get());
-				CommonActions->DeleteElements(ElementsToDelete.Get(), InWorld, SelectionSet, FTypedElementDeletionOptions());
-				ElementsToDelete->Reset();
+				if (ElementsToDelete->Num() > 0)
+				{
+					const FScopedTransaction Transaction(NSLOCTEXT("UnrealEd", "DeleteElements", "Delete Elements"));
+					CommonActions->DeleteElements(ElementsToDelete.Get(), InWorld, SelectionSet, FTypedElementDeletionOptions());
+					ElementsToDelete->Reset();
+				}
 			}
 		}
 		return true;
